+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # oEmbed Changelog
 
+## 1.0.1 - 2018-11-26
+
+### Updated
+- Changed preview controller action access to prevent anonymous access
+- Refactored the preview action to use a template with wrapper to allow for future styling and updates
+
 ## 1.0.0 - 2018-11-18
 
 ### Added